requirements.txt
absl-py==0.1.11
astor==0.6.2
audioread==2.1.5
bandmat==0.6
bleach==1.5.0
certifi==2018.1.18
Cython==0.27.3
decorator==4.2.1
-e git+https://github.com/edwinlimlx/deepvoice3_pytorch.git@cf3f476d0ab0f3984bd2e7d1d545f7867f73e295#egg=deepvoice3_pytorch
docopt==0.6.2
falcon==1.4.1
fastdtw==0.3.2
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
inflect==0.2.5
joblib==0.11
librosa==0.6.0
llvmlite==0.22.0
lws==1.0
Markdown==2.6.11
nltk==3.2.5
nnmnkwii==0.0.13
numba==0.37.0
numpy==1.14.1
protobuf==3.5.2
pysptk==0.1.10
python-mimeparse==1.6.0
resampy==0.2.0
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorboard==1.6.0
tensorboardX==1.1
tensorflow-gpu==1.6.0
termcolor==1.1.0
tqdm==4.19.6
Unidecode==1.0.22
Werkzeug==0.14.1
